Move script functions to happen inline and to not create new connections to the db, but to re-use existing ones
